# openapi.yaml
openapi: 3.0.3
info:
  title: Personal Budget Part II
  version: '1.0'
  description: >-
    Extend the Personal Budget API created in Personal Budget, Part I, to
    include persisting data
servers:
  - url: http://localhost:4001/envelopes
    variables: {}
paths:
...
  /{envId}/items/:
    summary: Get items in envelope
    get:
      summary: Get all items in envelope
      description: Returns an array of all items in an envelope
      operationId: get_items
      parameters:
        - name: envId
          in: path
          description: The id of the envelope
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                Items from envelope:
                  value: |-
                    [
                        {
                            "sub_id": 1,
                            "name": "Book",
                            "cost": "£6.00",
                            "category_id": 2
                        },
                        {
                            "sub_id": 2,
                            "name": "Newspaper",
                            "cost": "£10.00",
                            "category_id": 2
                        }
                    ]
        '400':
          description: BAD REQUEST
          content:
            text/plain; charset=utf-8:
              examples:
                Bad envelope id:
                  value: Envelope id should be a positive integer
                No such envelope:
                  value: No such envelope ID 1
        '500':
          description: Internal Server Error
          content:
            text/plain; charset=utf-8:
              examples:
                Internal server error:
                  value: connect ECONNREFUSED 127.0.0.1:5432
      tags:
        - Items
    post:
      summary: Creates new envelope
      description: Creates a new envelope and returns it
      operationId: post_item
      parameters:
        - name: envId
          in: path
          description: The id of the envelope
          required: true
          schema:
            type: integer
        - name: name
          in: query
          description: Name for the item
          required: true
          schema:
            type: string
        - name: cost
          in: query
          description: Cost of the item
          required: true
          schema:
            type: number
      responses:
        '201':
          description: Created
          content:
            application/json:
              examples:
                Envelope created:
                  value: |-
                    {
                        "sub_id": 3,
                        "name": "Bus ticket",
                        "cost": "£5.00",
                        "category_id": 1
                    }
        '400':
          description: Bad Request
          content:
            text/plain; charset=utf-8:
              examples:
                Bad envelope id:
                  value: Envelope id should be a positive integer
                Missing required parameters:
                  value: 'Missing required parameters: name, cost'
                No such envelope:
                  value: No such envelope ID 1
                Non-number cost value:
                  value: Cost should be a number
                Negative cost value:
                  value: Cost should be a positive number
                Cost value exceeds available budget:
                  value: Available budget in envelope is <value>
                  description: The new item's cost exceeds available budget in envelope
        '500':
          description: Internal Server Error
          content:
            text/plain; charset=utf-8:
              examples:
                Internal server error:
                  value: connect ECONNREFUSED 127.0.0.1:5432
      tags:
        - Items
  /{envId}/items/{itemId}:
    summary: Operations with items
    get:
      summary: Retrieve an item
      description: Retrieve an item by envelope id and item id
      operationId: get_item
      parameters:
        - name: envId
          in: path
          description: The id of the item's envelope
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          description: The id of the item
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ok
          content:
            application/json:
              examples:
                Item returned:
                  value: |-
                    {
                        "sub_id": 1,
                        "name": "Book",
                        "cost": "£6.00",
                        "category_id": 2
                    }
        '400':
          description: Bad Request
          content:
            text/plain; charset=utf-8:
              examples:
                Bad envelope id:
                  value: Envelope id should be a positive integer
                Bad item id:
                  value: Item id should be a positive integer
                No such envelope:
                  value: No such envelope ID 1
                No such item:
                  value: No such item
        '500':
          description: Internal Server Error
          content:
            text/plain; charset=utf-8:
              examples:
                Internal server error:
                  value: connect ECONNREFUSED 127.0.0.1:5432
      tags:
        - Items
    put:
      summary: Update all item's fields
      description: Updates an existing item and returns it
      operationId: put_item
      parameters:
        - name: envId
          in: path
          description: The id of the item's envelope
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          description: The id of the item to be updated
          required: true
          schema:
            type: integer
        - name: name
          in: query
          description: Name for the item
          required: true
          schema:
            type: string
        - name: cost
          in: query
          description: Cost of the item
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                Item updated:
                  value: |-
                    {
                        "sub_id": 1,
                        "name": "Book",
                        "cost": "£7.00",
                        "category_id": 2
                    }
        '400':
          description: Bad Request
          content:
            text/plain; charset=utf-8:
              examples:
                Bad envelope id:
                  value: Envelope id should be a positive integer
                Bad item id:
                  value: Item id should be a positive integer
                Missing required parameters:
                  value: 'Missing required parameters: name, cost'
                No such envelope:
                  value: No such envelope ID 1
                No such item:
                  value: No such item
                Non-number cost value:
                  value: The cost value has to be a number
                Negative cost value:
                  value: The cost value has to be a positive number
                Envelope name exists:
                  value: The envelope name has to be unique
                Cost value exceeds available budget:
                  value: New cost exceeds <value>
                  description: >-
                    The value of the item exceeds available budget in the
                    envelope
                No update:
                  value: Nothing to be updated
                  description: Supplied values already match the item's
        '500':
          description: Internal Server Error
          content:
            text/plain; charset=utf-8:
              examples:
                Internal server error:
                  value: connect ECONNREFUSED 127.0.0.1:5432
      tags:
        - Items
    patch:
      summary: Update an item's fields
      description: Updates an existing item and returns it
      operationId: patch_item
      parameters:
        - name: envId
          in: path
          description: The id of the item's envelope
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          description: The id of the item to be updated
          required: true
          schema:
            type: integer
        - name: name
          in: query
          description: Name for the item
          required: false
          schema:
            type: string
        - name: cost
          in: query
          description: Cost of the item
          required: false
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                Item updated:
                  value: |-
                    {
                        "sub_id": 1,
                        "name": "Book",
                        "cost": "£7.00",
                        "category_id": 2
                    }
        '400':
          description: Bad Request
          content:
            text/plain; charset=utf-8:
              examples:
                Bad envelope id:
                  value: Envelope id should be a positive integer
                Bad item id:
                  value: Item id should be a positive integer
                No valid parameters:
                  value: No valid parameters found (name/cost)
                  description: None of an item's parameters has been supplied
                No such envelope:
                  value: No such envelope ID 1
                No such item:
                  value: No such item
                Non-number cost value:
                  value: The cost value has to be a number
                Negative cost value:
                  value: The cost value has to be a positive number
                Cost value exceeds available budget:
                  value: New cost exceeds <value>
                  description: >-
                    The value of the item exceeds available budget in the
                    envelope
                No update:
                  value: Nothing to be updated
                  description: Supplied values already match the item's
        '500':
          description: Internal Server Error
          content:
            text/plain; charset=utf-8:
              examples:
                Internal server error:
                  value: connect ECONNREFUSED 127.0.0.1:5432
      tags:
        - Items
    delete:
      summary: Deletes an item
      description: Deletes an existing item
      operationId: delete_item
      parameters:
        - name: envId
          in: path
          description: The id of the item's envelope
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          description: The id of the item to be updated
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No content
        '400':
          description: BAD REQUEST
          content:
            text/plain; charset=utf-8:
              examples:
                Bad envelope id:
                  value: Envelope id should be a positive integer
                Bad item id:
                  value: Item id should be a positive integer
                No such envelope:
                  value: No such envelope ID 1
                No such item:
                  value: No such item
        '500':
          description: Internal Server Error
          content:
            text/plain; charset=utf-8:
              examples:
                Internal server error:
                  value: connect ECONNREFUSED 127.0.0.1:5432
      tags:
        - Items
  /{srcEnvId}/items/{itemId}/moveTo/{destEnvId}:
    summary: Move an item
    description: Move an item to a different envelope
    patch:
      summary: Move an item
      description: Move an item to a different envelope
      operationId: move_item
      parameters:
        - name: srcEnvId
          in: path
          description: The id of the item's envelope
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          description: The id of the item to be moved
          required: true
          schema:
            type: integer
        - name: destEnvId
          in: path
          description: The id of the destination envelope
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              examples:
                Updated item:
                  value: |-
                    {
                        "sub_id": 4,
                        "name": "Book",
                        "cost": "£7.00",
                        "category_id": 1
                    }
        '400':
          description: BAD REQUEST
          content:
            text/plain; charset=utf-8:
              examples:
                Bad source envelope id:
                  value: Envelope id should be a positive integer
                Bad destination envelope id:
                  value: Destination envelope id should be a positive integer
                Bad item id:
                  value: Item id should be a positive integer
                No such envelope:
                  value: No such envelope ID 1
                No such item:
                  value: No such item
                Not enough budget:
                  value: Not enough budget in destination envelope
                  description: Item's cost exceeds available budget in destination envelope
        '500':
          description: Internal Server Error
          content:
            text/plain; charset=utf-8:
              examples:
                Internal server error:
                  value: connect ECONNREFUSED 127.0.0.1:5432
      tags:
        - Items